requirements-celery.txt
# ============================================================
# Celery Worker Requirements
# ============================================================

# Core task queue
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# Message broker/result backend
hiredis==2.2.3  # Faster Redis parser

# Scheduled tasks
celery-beat==2.5.0
python-dateutil==2.8.2

# Monitoring
flower==2.0.1
prometheus-client==0.19.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# HTTP requests (for downloading images, calling services)
httpx==0.25.2
requests==2.31.0

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78

# ML inference (optional, depends on your service)
# torch==2.1.0
# transformers==4.35.0
# diffusers==0.21.0

# Cloud storage
boto3==1.29.7          # AWS S3
google-cloud-storage==2.10.0  # Google Cloud Storage

# Error tracking
sentry-sdk==1.39.1

# Logging & monitoring
python-json-logger==2.0.7

# Environment variables
python-dotenv==1.0.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0

# Development & debugging (optional)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# pytest-cov==4.1.0
